reorganize router
