+++ code-clash-frontend/src/subclasses/header/header.tsx
@@ -1,6 +1,6 @@
 // React
 import * as React from "react";
-import { Link } from "react-router-dom";
+import {BrowserRouter, Link} from "react-router-dom";
 // Style
 import './header.css';
 
@@ -14,28 +14,26 @@ class Header extends React.Component<any, any>
     render()
     {
         return (
-                    <div>
-                        <div className="header">
-                            <div className="header-link">
-                                <Link to="/"><h3>Code Clash</h3></Link>
-                            </div>
-                            <div className="header-link">
-                                <Link to="/blocks">Blocks</Link>
-                            </div>
-                            <div className="header-link">
-                                <Link to="/register">Register</Link>
-                            </div>
-                            <div className="header-link">
-                                <Link to="/signin">Sign In</Link>
-                            </div>
-                            <div className="header-link">
-                                <Link to="/account">Account</Link>
-                            </div>
-                            <div className="header-link">
-                                <Link to="/upload">Upload</Link>
-                            </div>
-                        </div>
-                    </div>
+            <div className="header">
+                <div className="header-link">
+                    <Link to="/"><h3>Code Clash</h3></Link>
+                </div>
+                <div className="header-link">
+                    <Link to="/blocks">Blocks</Link>
+                </div>
+                <div className="header-link">
+                    <Link to="/register">Register</Link>
+                </div>
+                <div className="header-link">
+                    <Link to="/signin">Sign In</Link>
+                </div>
+                <div className="header-link">
+                    <Link to="/account">Account</Link>
+                </div>
+                <div className="header-link">
+                    <Link to="/upload">Upload</Link>
+                </div>
+            </div>
         )
     }
 }
